updated profile fetch call
