+++ src/content/pages/Posting.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route } from 'react-router-dom'
 
-import newItem from './New/newItem'
-import newList from './New/newList'
+import NewItem from './New/NewItem'
+import NewList from './New/NewList'
 
 const Posting = props => {
 
@@ -12,10 +12,10 @@ const Posting = props => {
         //As a suggestion maybe we should have two different forms and use routers??
         <div>
             <div className="addForms">
-                <Route exact path='/newItem' component={newItem} /> 
+                <Route exact path='/newItem' component={NewItem} /> 
             </div>
             <div>
-                <Route exact path='/newListing' component={newList} />
+                <Route exact path='/newListing' component={NewList} />
                 
             </div>
         </div>
